+++ NEWS.md
@@ -1,3 +1,5 @@
+# labelled (development version)
+
 # labelled 2.6.0
 
 * `look_for()` has been redesigned: